Add files via upload
This website is about a videogame I played recently, Hogwarts Legacy. I started website developing on the account of my friend's encouragement, as he is an expert in this field. I didn't regret it, because frontend development became the part of my circle of interest, even though my HTML and CSS skills are far from perfect. As you can see, my website barely has any user-friendly functions and the visual representation is also pretty weak, but I keep upgrading it in order to make it a decent website for today's standarts.